Making some preliminary classes (Much to be done, this is just a basic start)
My vision is that we have class objects that contain the results of stuff we retrieve from the database. That's why the variables are lists. In addition, we can also use this to send info to the database.

Functions for the actual queries themselves should probably go outside the classes (at least the way I have them currently). The classes are meant to do simple functions and be used later in the actual queries our app needs to perform